Fix: null error
1. 홈페이지 들어가면 발생하는 null 에러 해결
 - 세션 Class의 추천 칼로리에 .toInt(0메소드 제거
 - 세션에서 Int 값으로 변경하여 받아옴

2. 섭취량 인디게이터 두께 변경
 - 기존 5픽셀에서 10픽셀로 조정

3. main함수 로직 refactoring

4. 로딩화면 개선
 - 기존 로딩화면 제거 후 새로운 인디게이터 적용
 (blur mask 사용위해 github 코드 사용, 추후 수정요망)

5. 이미지 선택 시 뒤로가기 버튼 누르면 에러 뜨는 현상 수정

+++ lib/session.dart
@@ -13,7 +13,7 @@ class Session {
   void setInfo(userInfo) {
     _userInfo = userInfo;
 
-    _dietInfo['recom_cal'] = ((_userInfo['height'] - 100) * 0.9) * 30;
+    _dietInfo['recom_cal'] = (((_userInfo['height'] - 100) * 0.9) * 30).toInt();
     _dietInfo['recom_hydrate'] = ((_dietInfo['recom_cal'] * 0.55) / 4).toInt();
     _dietInfo['recom_protein'] = ((_dietInfo['recom_cal'] * 0.2) / 4).toInt();
     _dietInfo['recom_fat'] = ((_dietInfo['recom_cal'] * 0.25) / 9).toInt();
